requirements.txt
requests
python-dotenv
openai[agents] 